Add small and large screen
