+++ shared/src/commonMain/kotlin/details/RecipeDetailsLarge.kt
@@ -1,0 +1,302 @@
+import androidx.compose.animation.core.animateIntOffsetAsState
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.nestedscroll.NestedScrollConnection
+import androidx.compose.ui.input.nestedscroll.NestedScrollSource
+import androidx.compose.ui.input.nestedscroll.nestedScroll
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.Velocity
+import androidx.compose.ui.unit.dp
+import details.StepsAndDetails
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import model.Recipe
+import org.jetbrains.compose.resources.ExperimentalResourceApi
+import org.jetbrains.compose.resources.resource
+import sensor.Listener
+import sensor.SensorData
+import sensor.SensorManager
+import sharedelementtransaction.SharedMaterialContainer
+import kotlin.math.PI
+
+
+@OptIn(ExperimentalResourceApi::class)
+@Composable
+fun RecipeDetailsLarge(
+    recipe: Recipe,
+    imageBitmap: ImageBitmap,
+    chefImage: ImageBitmap?,
+    goBack: () -> Unit,
+    sensorManager: SensorManager,
+) {
+    val backgroundImage = remember { mutableStateOf<ImageBitmap?>(null) }
+    val imageRotation = remember { mutableStateOf(0) }
+    val sensorDataLive = remember { mutableStateOf(SensorData(0.0f, 0.0f)) }
+    val roll by derivedStateOf { (sensorDataLive.value.roll * 20).coerceIn(-2f, 2f) }
+    val pitch by derivedStateOf { (sensorDataLive.value.pitch * 20).coerceIn(-2f, 2f) }
+
+    val tweenDuration = 300
+
+    sensorManager.registerListener(object : Listener {
+        override fun onUpdate(sensorData: SensorData) {
+            sensorDataLive.value = sensorData
+        }
+    })
+
+    val animatedOffset = animateIntOffsetAsState(
+        targetValue = IntOffset((roll * 6f).toInt(), -(pitch * 6f).toInt()),
+        animationSpec = tween(tweenDuration)
+    )
+    val animatedOffset2 = animateIntOffsetAsState(
+        targetValue = IntOffset(-roll.toInt(), pitch.toInt()),
+        animationSpec = tween(tweenDuration)
+    )
+
+    LaunchedEffect(Unit) {
+        withContext(Dispatchers.Unconfined) {
+            if (recipe.bgImageName.isNotEmpty()) {
+                backgroundImage.value = resource(recipe.bgImageName).readBytes().toImageBitmap()
+            }
+        }
+    }
+
+    val nestedScrollConnection = remember {
+        object : NestedScrollConnection {
+            override fun onPreScroll(
+                available: Offset,
+                source: NestedScrollSource
+            ): Offset {
+                imageRotation.value += (available.y * 0.5).toInt()
+                return Offset.Zero
+            }
+
+            override fun onPostScroll(
+                consumed: Offset,
+                available: Offset,
+                source: NestedScrollSource
+            ): Offset {
+                val delta = available.y
+                imageRotation.value += ((delta * PI / 180) * 10).toInt()
+                return super.onPostScroll(consumed, available, source)
+            }
+
+            override suspend fun onPreFling(available: Velocity): Velocity {
+                imageRotation.value += available.y.toInt()
+                return super.onPreFling(available)
+            }
+        }
+    }
+
+
+    Box(modifier = Modifier.fillMaxSize()) {
+        Row(modifier = Modifier.background(sugar)) {
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .weight(1f)
+            ) {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .background(
+                            Color.Transparent,
+                            RoundedCornerShape(topEnd = 35.dp, bottomEnd = 35.dp)
+                        ).clip(
+                            RoundedCornerShape(topEnd = 35.dp, bottomEnd = 35.dp)
+                        ).fillMaxHeight(),
+                ) {
+                    SharedMaterialContainer(
+                        key = "$recipe ",
+                        screenKey = DetailsScreen,
+                        color = recipe.bgColor,
+                        shape = RoundedCornerShape(topEnd = 35.dp, bottomEnd = 35.dp),
+                        transitionSpec = MaterialFadeInTransitionSpec
+                    ) {
+
+                        // background image + its shadow
+                        Box(modifier = Modifier.fillMaxSize()) {
+                            backgroundImage.value?.let {
+                                Image(
+                                    bitmap = blurFilter(it, getPlatformContext()),
+                                    contentDescription = null,
+                                    contentScale = ContentScale.Fit,
+                                    modifier = Modifier
+                                        .offset {
+                                            animatedOffset.value
+                                        }.graphicsLayer(
+                                            scaleX = 1.050f,
+                                            scaleY = 1.050f
+                                        ),
+                                    colorFilter = ColorFilter.tint(
+                                        orangeDark.copy(alpha = 0.3f)
+                                    )
+                                )
+                                Image(
+                                    bitmap = it,
+                                    contentDescription = null,
+                                    contentScale = ContentScale.Fit,
+                                    modifier = Modifier.background(
+                                        Color.Transparent,
+                                        RoundedCornerShape(
+                                            bottomEnd = 35.dp,
+                                            bottomStart = 35.dp
+                                        ),
+                                    ).offset {
+                                        animatedOffset2.value
+
+                                    }.graphicsLayer(
+                                        shadowElevation = 8f,
+                                        scaleX = 1.050f,
+                                        scaleY = 1.050f
+                                    ),
+                                )
+                            }
+
+
+                            // image shadows and image
+                            Box(
+                                modifier = Modifier.aspectRatio(1f).padding(32.dp)
+                                    .align(Alignment.Center)
+                            ) {
+                                SharedMaterialContainer(
+                                    key = recipe.image,
+                                    screenKey = "DetailsScreen",
+                                    color = Color.Transparent,
+                                    transitionSpec = FadeOutTransitionSpec
+                                ) {
+                                    Box(modifier = Modifier.padding(32.dp)) {
+                                        Box(
+                                            modifier = Modifier
+                                                .offset {
+                                                    IntOffset(
+                                                        x = (roll * 2).dp.roundToPx(),
+                                                        y = -(pitch * 2).dp.roundToPx()
+                                                    )
+                                                }
+                                        ) {
+
+                                            Image(
+                                                bitmap = imageBitmap,
+                                                contentDescription = null,
+                                                modifier = Modifier.aspectRatio(1f)
+                                                    .align(Alignment.Center)
+                                                    .padding(16.dp)
+                                                    .shadow(
+                                                        elevation = 16.dp,
+                                                        shape = CircleShape,
+                                                        clip = false,
+                                                        ambientColor = Color.Red,
+                                                        spotColor = Color.Red,
+                                                    ),
+                                                colorFilter = ColorFilter.tint(
+                                                    orangeDark.copy(alpha = 0.0f)
+                                                )
+                                            )
+                                        }
+
+                                        Image(
+                                            bitmap = imageBitmap,
+                                            contentDescription = null,
+                                            modifier = Modifier.aspectRatio(1f)
+                                                .align(Alignment.Center)
+                                                .padding(16.dp)
+                                                .rotate(imageRotation.value.toFloat())
+                                                .background(
+                                                    Color.Transparent,
+                                                    CircleShape,
+                                                )
+                                        )
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+
+                Box(
+                    modifier = Modifier.padding(32.dp)
+                        .clip(RoundedCornerShape(50))
+                        .clickable {
+                            goBack()
+                        }.background(
+                            color = Color.White,
+                            shape = RoundedCornerShape(50)
+                        ).padding(top = 8.dp, bottom = 8.dp, start = 16.dp, end = 16.dp)
+                ) {
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Icon(
+                            imageVector = Icons.Default.ArrowBack,
+                            contentDescription = null,
+                            tint = Color.Black,
+                            modifier = Modifier.size(20.dp)
+                        )
+                        Spacer(Modifier.padding(start = 8.dp))
+                        Text(text = "Back to Recipes", style = MaterialTheme.typography.subtitle2)
+                    }
+                }
+            }
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .weight(1f)
+
+            ) {
+
+                val listState = rememberLazyListState()
+                Box(
+                    modifier = Modifier.fillMaxSize().background(color = sugar)
+                ) {
+                    LazyColumn(
+                        contentPadding = PaddingValues(32.dp),
+                        userScrollEnabled = true,
+                        verticalArrangement = Arrangement.Absolute.spacedBy(16.dp),
+                        modifier = Modifier.fillMaxSize().nestedScroll(nestedScrollConnection),
+                        state = listState
+                    ) {
+                        StepsAndDetails(recipe, chefImage)
+                    }
+                }
+            }
+        }
+    }
+}
