+++ shared/src/commonMain/kotlin/recipeslist/RecipeImage.kt
@@ -11,7 +11,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.unit.dp
 
 /**
  * Created by abdulbasit on 20/06/2023.
