+++ shared/src/commonMain/kotlin/recipeslist/RecipesList.kt
@@ -2,10 +2,11 @@ package recipeslist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyGridState
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -13,8 +14,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.unit.dp
 import model.Recipe
 import sugar
 
@@ -27,16 +28,24 @@ import sugar
 fun RecipesListScreen(
     items: List<Recipe>, width: Int,
     updateIds: String,
-    onClick: (recipe: Recipe, offset: Offset, size: Int, imageBitmap: ImageBitmap) -> Unit,
+    onClick: (recipe: Recipe, imageBitmap: ImageBitmap) -> Unit,
+    isLarge: Boolean,
 ) {
-    val listState = rememberLazyListState()
-    LazyColumn(modifier = Modifier.fillMaxSize().background(sugar), state = listState) {
-        items(items) { item ->
+    val listState = rememberLazyGridState()
+    LazyVerticalGrid(
+        modifier = Modifier.fillMaxSize().background(sugar).padding(16.dp),
+        state = listState,
+        columns = GridCells.Fixed(if (isLarge) 3 else 1)
+    ) {
+        items(items.size) { item ->
             RecipeListItemWrapper(
                 scrollDirection = listState.isScrollingUp(),
                 child = {
                     RecipeListItem(
-                        recipe = item, width = width, onClick = onClick, updateIds = updateIds
+                        recipe = items[item],
+                        width = width,
+                        onClick = onClick,
+                        updateIds = updateIds
                     )
                 },
             )
@@ -45,7 +54,7 @@ fun RecipesListScreen(
 }
 
 @Composable
-private fun LazyListState.isScrollingUp(): Boolean {
+private fun LazyGridState.isScrollingUp(): Boolean {
     var previousIndex by remember(this) { mutableStateOf(firstVisibleItemIndex) }
     var previousScrollOffset by remember(this) { mutableStateOf(firstVisibleItemScrollOffset) }
     return remember(this) {
